requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
fastapi==0.87.0
fastapi-csrf-protect==0.2.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
motor==3.1.1
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
python-decouple==3.6
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.19.0
